add check for profileproperty.has_language to export
The export was looking at whether each statement had a language attribute.
This isn’t a problem until you change the language requirement in the profile,
and then the statement is discarded because it still has a language, but the
language is no longer used. This change ignores the language attribute, even
if there is one in the statement, if the property doesn’t require a language.